feature: merge from master

+++ server/apps/node_mgmt/views/package.py
@@ -14,7 +14,8 @@ class PackageMgmtView(
     mixins.CreateModelMixin,
     mixins.DestroyModelMixin,
     mixins.ListModelMixin,
-    GenericViewSet):
+    GenericViewSet,
+):
     queryset = PackageVersion.objects.all()
     serializer_class = PackageVersionSerializer
     filterset_class = PackageVersionFilter
@@ -30,24 +31,23 @@ class PackageMgmtView(
         return super().destroy(request, *args, **kwargs)
 
     def create(self, request, *args, **kwargs):
-        uploaded_file = request.FILES.get('file')
+        uploaded_file = request.FILES.get("file")
         if not uploaded_file:
             return WebUtils.response_error(error_message="请上传文件")
 
-        package_type = request.data.get('type')  # collector 或 controller
-        os_type = request.data.get('os')  # linux 或 windows
-        object_name = request.data.get('object')  # 采集器/控制器名称
+        package_type = request.data.get("type")  # collector 或 controller
+        os_type = request.data.get("os")  # linux 或 windows
+        object_name = request.data.get("object")  # 采集器/控制器名称
 
         # 校验必填参数
         if not all([package_type, os_type, object_name]):
-            return WebUtils.response_error(error_message="请填写完整的包类型、操作系统和对象名称")
+            return WebUtils.response_error(
+                error_message="请填写完整的包类型、操作系统和对象名称"
+            )
 
         # 校验包并自动识别版本
         is_valid, error_message, parsed_info = PackageService.validate_package(
-            uploaded_file.name,
-            package_type,
-            os_type,
-            object_name
+            uploaded_file.name, package_type, os_type, object_name
         )
 
         if not is_valid:
@@ -58,20 +58,33 @@ class PackageMgmtView(
             os=os_type,
             type=package_type,
             object=object_name,
-            version=parsed_info['version'],  # 自动识别的版本号
-            name=parsed_info['name_without_version'],  # 存储去掉版本号的文件名
-            description=request.data.get('description', ''),
+            version=parsed_info["version"],  # 自动识别的版本号
+            name=parsed_info["name_without_version"],  # 存储去掉版本号的文件名
+            description=request.data.get("description", ""),
             created_by=request.user.username,
             updated_by=request.user.username,
         )
 
+        existing_package = parsed_info.get("existing_package")
+        if existing_package:
+            PackageService.upload_file(uploaded_file, data)
+            existing_package.description = data.get(
+                "description", existing_package.description
+            )
+            existing_package.updated_by = request.user.username
+            existing_package.save(
+                update_fields=["description", "updated_by", "updated_at"]
+            )
+            return WebUtils.response_success(
+                PackageVersionSerializer(existing_package).data
+            )
+
         serializer = self.get_serializer(data=data)
         serializer.is_valid(raise_exception=True)
 
-        # 上传文件，成功了再保存数据
         PackageService.upload_file(uploaded_file, data)
         self.perform_create(serializer)
-        
+
         return WebUtils.response_success(serializer.data)
 
     @action(detail=False, methods=["get"], url_path="download/(?P<pk>.+?)")
